requirements.txt
opencv-python
scikit-learn
pandas
matplotlib
numpy
pillow
seaborn
scipy
keras
torchvision
tensorflow
h5py
Augmentor
tqdm
protobuf==3.20
google-generativeai


